==> STEP_1_National_Parks.ipynb <==
## Step 1 - Starting off with just the zip codes of all the national parks. I'll add in more data as we go along

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.cm as cm
import matplotlib.patches as mpatches
from mpl_toolkits.basemap import Basemap
import zipcode
%matplotlib inline 

master = pd.DataFrame.from_csv("nps_all.csv")
master.head(7)

## This csv just contains the names and zipcodes of all the national parks. The zip codes need to be formatted to zip code format, as the leading zeroes have fallen off. The names of all the national parks were pulled from https://irma.nps.gov/Stats/SSRSReports/National%20Reports/Annual%20Recreation%20Visitation%20By%20Park%20(1979%20-%20Last%20Calendar%20Year). Zipcodes are sourced from google. 

#this function converts month numbers to month strings
def num_month(month):
    months = ['JAN', 'FEB', 'MAR', 'APR', 'MAY', 'JUN', 'JUL', 'AUG', 'SEP', 'OCT', 'NOV', 'DEC']
    return months.index(month) + 1

files = master.index

#joining up each park's data with the master table
for i in range(0, len(files)):
    #getting the file
    part = pd.DataFrame.from_csv("tables/" + files[i] + ".csv")
    part = part.drop('Textbox5', 1)
    avgs = pd.DataFrame(part.ix[0].astype(float))
    #ranking the averages
    avgs = avgs.sort(columns='Average', ascending=False)
    ranking = avgs.index
    best = num_month(ranking[0])
    best2 = num_month(ranking[1])
    best3 = num_month(ranking[2])
    best4 = num_month(ranking[3])
    worst = num_month(ranking[-1])
    #putting the best, 2nd best, so on months into the master 
    master.ix[i,1] = best
    master.ix[i,2] = best2
    master.ix[i,3] = best3
    master.ix[i,4] = best4
    master.ix[i,5] = worst
    
master.head()
    

## All of the visitor information files are from https://irma.nps.gov/Stats/SSRSReports/Park%20Specific%20Reports/Recreation%20Visitors%20By%20Month%20(1979%20-%20Last%20Calendar%20Year)?Park=ACAD. 
## They were reformatted using Excel VBA to include averages. To find the "best season", and "2nd best season" and so on, I averaged each month from 2000 to 2015 for every national park. The month with the highest average is marked the "best season."

#this function translate the best month to a corresponding color
def color_translate(color_num):
    colors = ['red','coral','orange','gold','yellow','greenyellow','green','teal','blue','slateblue','blueviolet','darkmagenta']
    return colors[color_num - 1]

color_translate(int(master.ix[0,1]))

#using the above function to assign a color to each park
master['Best color'] = master['Best Season'].astype(int)
master['Best color'] = master['Best color'].map(color_translate)
master.head()

## Now each park has a color assigned to it that corresponds to its most popular season. I'll use this later when we plot the parks.

#adding in leading zeroes to zip codes and formatting them as strings
master['Zip'] = master['Zip'].astype(str)
master['Zip'] = master['Zip'].map(lambda x: x.rjust(5, "0"))

master.head()

#this function returns the latitude of a zipcode object
def make_lat(zobject):
    if zobject is not None:
        return zobject.lat
    else:
        return 0.0

#this function returns the longitute of a zipcode object
def make_lon(zobject):
    if zobject is not None:
        return zobject.lon
    else:
        return 0.0


#using the above functions to put latitudes and longitudes into the master
master['ZObjects'] = master['Zip'].map(lambda x: zipcode.isequal(x))
master['Lat'] = master['ZObjects'].map(lambda x: make_lat(x))
master['Lon'] = master['ZObjects'].map(lambda x: make_lon(x))
master.head()

## Almost all of the parks have latitude and longitude data now, and I'll use that when I plot the parks. Some zipcodes however are not available in the zipcode library, so I'll save this as it is and find the zipcodes on my own. 

#checking to see which zipcodes are not available in the zipcode library. 
master[(master.Lat == 0.0) & (master.Lon == 0.0)]

master.to_csv("Master.csv")
