==> STEP_3_Regexes.ipynb <==
## Step 3 - Getting weather station data and reformatting it with regexes

import pandas as pd
import re

## data from https://www.ncdc.noaa.gov/data-access/land-based-station-data/land-based-datasets/climate-normals/1981-2010-normals-data

#importing the csv
old_csv = pd.DataFrame.from_csv("weather/station_zips.csv", index_col=None, header=None)

#using a regex to separate each line appropriately
for i in range(len(old_csv)):
    line = old_csv.ix[i,0]
    separated = re.match(r'^([\w]{11}) ([\d]{5}) ([A-za-z -]+)$', line)
    parts = list(separated.groups())
    old_csv.ix[i,0] = parts[0]
    old_csv.ix[i,1] = parts[1]
    old_csv.ix[i,2] = parts[2]


old_csv

#saving the changes
old_csv.to_csv("weather/station_zips.csv")

#importing the csv
old_csv = pd.DataFrame.from_csv("weather/stations_locations.csv", index_col=None, header=None)

#using a regex to separate each line appropriately
for i in range(len(old_csv)):
    line = old_csv.ix[i,0]
    separated = re.match(r'^([\w]{11})[ ]+([\d\.-]+)[ ]+([\d\.-]+)[ ]+([\d\.-]+)[ ]+([A-Z]{2})[ ]+([A-Z\d -/&]+?)[ ]*([GSN]?[HCN]?)[ ]*([\d]{0,5})$', line)
    parts = list(separated.groups())
    old_csv.ix[i,0] = parts[0]
    old_csv.ix[i,1] = parts[1]
    old_csv.ix[i,2] = parts[2]
    old_csv.ix[i,3] = parts[3]
    old_csv.ix[i,4] = parts[4]
    old_csv.ix[i,5] = parts[5]
    old_csv.ix[i,6] = parts[6]
    old_csv.ix[i,7] = parts[7]
    

old_csv

#saving the changes
old_csv.to_csv("weather/stations_locations.csv")
