==> STEP_4_More_Regexes.ipynb <==
## Step 4 - Getting weather (temperature and precipitation) data and reformatting it with regexes

import pandas as pd

## According to the readme from https://www.ncdc.noaa.gov/data-access/land-based-station-data/land-based-datasets/climate-normals/1981-2010-normals-data temperatures are measured down to the tenth and precipitation numbers down to the hundreth.

#importing the csvs
monthly_temp_normals = pd.DataFrame.from_csv("weather/tavg.csv", index_col=None, header=None)
monthly_precip_normals = pd.DataFrame.from_csv("weather/prcp.csv", index_col=None, header=None)

#dropping Nan columns
monthly_precip_normals.drop(13,axis=1, inplace=True)
monthly_precip_normals.drop(14,axis=1, inplace=True)
monthly_precip_normals.drop(15,axis=1, inplace=True)
monthly_temp_normals.drop(13,axis=1, inplace=True)
monthly_temp_normals.drop(14,axis=1, inplace=True)
monthly_temp_normals.drop(15,axis=1, inplace=True)

monthly_precip_normals.head()

#for monthly precipitation, applying appropriate transformations
for i in range(len(monthly_precip_normals)):
    for j in range(1,13):
        precip = monthly_precip_normals.ix[i,j]
        precip = precip[0:-1]
        if precip == -7777:
            precip = 0
        else:
            precip = float(precip)/100
        monthly_precip_normals.ix[i,j] = precip


#for monthly temp, applying appropriate transformations
for i in range(len(monthly_temp_normals)):
    for j in range(1,13):
        temp = monthly_temp_normals.ix[i,j]
        temp = temp[0:-1]
        if temp == -7777:
            temp = 0
        else:
            temp = float(temp)/10
        monthly_temp_normals.ix[i,j] = temp
        

#saving these changes
monthly_precip_normals.to_csv("prcp.csv")
monthly_temp_normals.to_csv("tavg.csv")
