==> STEP_5_Joining_Data.ipynb <==
## Step 5 - Linking the closest weather station to each park by ID

import pandas as pd

### Since there is a lot of data, first I'll find the closest station to each place according to the latitude and longitude

stations = pd.DataFrame.from_csv("weather/stations_locations.csv")
stations.head()

parks = pd.DataFrame.from_csv("Ultimate_Master.csv")
parks.head()

parks['Closest Station'] = ""


for i in range(len(parks)):
    closest_dist = 10000
    closest_id = ""
    for j in range(len(stations)):
        dist = ((parks.ix[i,8] - stations.ix[j,1])**2 +(parks.ix[i,9] - stations.ix[j,2])**2)**(.5)
        if dist < closest_dist:
            closest_dist = dist
            closest_id = stations.ix[j,0]
    parks.ix[i,10] = closest_id

parks.to_csv("Ultimate_Master_2.csv")
