==> STEP_6_Joining_More_Data.ipynb <==
## Step 6 - Adding in Weather Data

import pandas as pd

parks = pd.DataFrame.from_csv("Ultimate_Master_2.csv")
temp = pd.DataFrame.from_csv("weather/tavg.csv")
precip = pd.DataFrame.from_csv("weather/prcp.csv")

#finding the best season in terms of the least precipitation
#and finding the best season in terms of temperature that is closest to 70 degrees Fahrenheit as that seems an ideal temp

parks['Best Precip'] = 0
parks['Best Temp'] = 0

for i in range(len(parks)):
    station = parks.ix[i,10]
    if station in precip:
        precip_season = precip[station].idxmin(axis=0)
        parks.ix[i,11] = precip_season
    if station in temp:
        temp_line = temp[station]
        temp_best = 0
        temp_season = 0
        for j in range(1,13):
            poss_temp = temp_line[j]
            if abs(70-poss_temp) < abs(70-temp_best):
                temp_best = poss_temp
                temp_season = j
        parks.ix[i,12] = temp_season
    
    
parks

# temp and precipitation data can't be found for these, i'll just keep using the visitor data for them
parks[parks['Best Precip'] == 0]

parks.to_csv("Ultimate_Master_3.csv")
